refactor(codegen): separate lowering concerns
