+++ crates/codegen/src/lower/expr.rs
@@ -291,10 +291,12 @@ impl<'gcx> Lowerer<'gcx> {
                             if let Some(selector) = self.lower_resolved_function_selector(base) {
                                 return builder.imm_u256(U256::from(selector) << 224);
                             }
-                            if let ExprKind::Member(receiver, function_name) = &base.kind {
-                                let selector =
-                                    self.compute_member_selector(receiver, *function_name);
-                                return builder.imm_u256(U256::from(selector) << 224);
+                            if let ExprKind::Member(_, function_name) = &base.kind {
+                                return self.err_value(
+                                    builder,
+                                    function_name.span,
+                                    format!("cannot resolve member function `.{function_name}`"),
+                                );
                             }
                             if let Some(selector) = self.ident_function_selector(base) {
                                 return builder.imm_u256(U256::from(selector) << 224);
@@ -963,7 +965,7 @@ impl<'gcx> Lowerer<'gcx> {
             return None;
         }
 
-        let field_tys = self.gcx.struct_field_types(struct_id).to_vec();
+        let field_tys = self.gcx.struct_field_types(struct_id);
         if field_tys.len() < MIN_BULK_ZERO_STRUCT_FIELDS {
             return None;
         }
@@ -2303,7 +2305,7 @@ impl<'gcx> Lowerer<'gcx> {
         let struct_size = (num_fields as u64) * 32;
         let struct_ptr =
             self.allocate_memory_object(builder, struct_size, crate::mir::MemoryObjectKind::Struct);
-        let field_tys = self.gcx.struct_field_types(struct_id).to_vec();
+        let field_tys = self.gcx.struct_field_types(struct_id);
         let arg_exprs =
             match self.ordered_args_for(args, Some(CallableParamSource::Struct(struct_id))) {
                 Ok(exprs) => exprs,
@@ -2311,11 +2313,11 @@ impl<'gcx> Lowerer<'gcx> {
             };
 
         // Store each argument into the corresponding field
-        for (i, (arg, &field_ty)) in arg_exprs.into_iter().zip(&field_tys).enumerate() {
+        for (i, (arg, &field_ty)) in arg_exprs.into_iter().zip(field_tys).enumerate() {
             // Memory struct fields hold memory values. Calldata reference
             // values therefore materialize recursively before storing their
             // pointer in the field slot.
-            let field_val = self.lower_return_value_for_ty(builder, arg, field_ty);
+            let field_val = self.coerce_value_for_type(builder, arg, field_ty);
             let field_addr = builder.memory_object_field_addr(
                 struct_ptr,
                 crate::mir::MemoryObjectLayout::structure(num_fields as u64),
